requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
discord.py==2.3.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
mysqlclient==2.2.1
SQLAlchemy==2.0.23
typing_extensions==4.9.0
yarl==1.9.4
